% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassVal.R
\name{ClassVal}
\alias{ClassVal}
\title{Validate a Tree Prediction with TreeCrown Segments}
\usage{
ClassVal(pred, rsp, rsp_class = NULL, reclass = NULL)
}
\arguments{
\item{pred}{RasterLayer - Prediction to be validated}

\item{rsp}{PolygonLayer - Response Polygons for a class}

\item{rsp_class}{numeric - The class representing the response area}

\item{reclass}{optional - numeric value of class to merge into response class, for multiple classes to merge use reclass=c(x,y)}
}
\value{
returns the validation score
\itemize{
\item return and plot the "tree \link{1} and no tree \link{2}" layer, will be reclassified if 'reclass' is used.
\item result table
}
\itemize{
\item nclass - amount of cells for selected class (may differ if 'reclass' is used)
\item nseg - total amount of cells for response area
\item nover - amount of cells for classification outside response area
\item nunder -  amount of cells inside response area without correct class
\item nhit - amount of cells inside response area with correct class
\item hitrate - response area with correct class in \% (sum of missrate and hitrate is 100\%)
\item missrate - response area without correct class in \% (sum of missrate and hitrate is 100\%)
\item overrate - amount of cells for responding class outside responding area in \% of total amount of cell of the class
}
}
\description{
calculates the amount of cells for a class which overlap with responding Polygons.
}
\author{
Andreas Schönberg
}
