% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IKARUS - Package.R
\docType{package}
\name{IKARUS - package}
\alias{IKARUS - package}
\title{Tool for Pixel-based Classifications}
\description{
In general the \code{IKARUS} package is a wrapper for the \code{CAST} package (Hanna Meyer,2020).
\code{IKARUS} is designed to easily perform pixel-based classifications with \code{CAST} and provides additional functions to handle input data.
}
\details{
IKARUS Pixel-based Classification Tool
}
\note{

}
\author{
Andreas Schönberg
}
\keyword{package}
