% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IKARUS - Package.R
\docType{data}
\name{lau_Stk - data}
\alias{lau_Stk - data}
\title{A RasterStack with spectral data}
\format{
\code{"raster::stack"}
}
\description{
An RGB orthophoto merged with the NIR band of a IRC image.
Additional artifically layers: NDVI along with Sum and Sobal filter of 3x3 for the NDVI.
The scene shows some trees in the Lautaret vally in the frence alps.
Bandnames are "blue","green","red","nir","NDVI","NDVI_sum3","NDVI_sobel3". Resolution 0.15 meter.
}
