% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IKARUS - Package.R
\docType{data}
\name{lau_TPoints - data}
\alias{lau_TPoints - data}
\title{Point layer with training areas.}
\format{
\code{"rgdal::readOGR"}
}
\description{
Point layer with training areas for 'lau_Stk'.
}
