% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IKARUS - Package.R
\docType{data}
\name{lau_TrainPoly - data}
\alias{lau_TrainPoly - data}
\title{Polygon layer with training areas.}
\format{
\code{"rgdal::readOGR"}
}
\description{
Polygon layer with training areas for 'lau_Stk'.
}
