% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFclass.R
\name{RFclass}
\alias{RFclass}
\title{Random Forest Classification with Random Cross-Validation}
\usage{
RFclass(
  tDat,
  predCol = "default",
  predStk = NULL,
  classCol = "class",
  Cores = 1
)
}
\arguments{
\item{tDat}{data.frame - with values of the predictors (see details)}

\item{predCol}{numeric - seq of columns with predictor values. By default uses 1:(length(tDat)-1) for tDat format computed by IKARUS::exrct_Traindat}

\item{predStk}{\itemize{
\item RasterStack - with the predictors.
}}

\item{classCol}{character - name of the column containing the class information}

\item{Cores}{numeric - amount of Cores to exclude from calculation, default = 1}
}
\value{
returns a list with the model and the prediction
}
\description{
RF Classification with CV
}
\details{
\itemize{
\item predCol -  specific predictors can be selected by setting predCol = x:y
\item tDat - the use of IKARUS::exrct_Traindat is recommended.
\item parallel processing - the function uses parallel processing for multicore processors. by default all cores -1 are used.
}
}
\examples{
# load data
require(caret)
require(CAST)
require(doParallel)
require(raster)
require(IKARUS)
lau_Stk <- raster::stack(system.file("extdata","lau_RGB.grd",package = "IKARUS"))
lau_tP <-rgdal::readOGR(system.file("extdata","lau_TrainPolygon.shp",package = "IKARUS"))
# handle CRS string
crs(lau_tP) <- crs(lau_Stk)
### extract values using 'exrct_Tdat' to generate training dataset
tDat <- exrct_Traindat(lau_tP,lau_Stk,"class")
# check for class column and predictor columns in input training dataset
head(tDat)
# classification
model1 <- RFclass(tDat = tDat,predCol = "default",predStk = lau_Stk,classCol = "class")
# check model
model1$model_cv

# plot prediction
plot(model1$prediction)
}
\author{
Andreas Schönberg
}
