% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exrct_Traindat.R
\name{exrct_Traindat}
\alias{exrct_Traindat}
\title{Extract training data from RasterStacks}
\usage{
exrct_Traindat(trainPoly, predStk, classCol = NULL, lyrname = names(predStk))
}
\arguments{
\item{trainPoly}{SpatialPolygonsDataframe -  with training polygons to assign values}

\item{predStk}{RasterStack - with layers to extract values from}

\item{classCol}{character - name of the column containing the class information, default=NULL}
}
\value{
Returns a data.frame with values of each Rasterlayer per pixel for the training polygons.
Additionally adds a column of the respective class information.
}
\description{
Extracts values from a RasterStack of predictor variables by training polygons
}
\details{
This function is used to extract training data from a Raster Stack. This training dataset is used for IKARUS::BestPredFFS and IKARUS::RFclass.
\itemize{
\item classCol - the column with information about the class of a polygon. Supports either character or numeric values for classes (eg 1,2,3 or "tree","stone","grass")
\item predictor selection - specific predictors can be selected by using [[]] in parameter 'predStk'. E.g. predStk = x[\link{1:4}].
}
}
\note{
the function will check for INF and or NA values. INF values are first set to NA and further all NA will be deleted to prevent errors in further processing with IKARUS::BestPredFFS and IKARUS::RFclass.
}
\examples{
# load data
require(raster)
require(IKARUS)
lau_Stk <- raster::stack(system.file("extdata","lau_Stk.tif",package = "IKARUS"))
#set layer names
names(lau_Stk)<- c("blue","green","red","nir","NDVI","NDVI_sum3","NDVI_sobel3")
lau_tP <-rgdal::readOGR(system.file("extdata","lau_TrainPoly.shp",package = "IKARUS"))
# handle CRS string
crs(lau_tP) <- crs(lau_Stk)
### check column names
names(lau_tP)
# -> lau_tP has both character and numeric class information
### extract values using character class information
tDat <- exrct_Traindat(lau_tP,lau_Stk,"class")
head(tDat)
### extract values using numeric class information
tDat2 <- exrct_Traindat(lau_tP,lau_Stk,"class_num")
head(tDat2)
}
\author{
Andreas Schönberg
}
